pyproject.toml
[tool.poetry]
name = "data-engineering"
version = "0.1.0"
description = ""
authors = ["Sebastián Dalceggio <sdalceggio@gmail.com>"]
readme = "README.md"
packages = [{include = "data_engineering"}]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.3"
requests = "^2.31.0"
sqlalchemy = "^2.0.19"
google-cloud-storage = "^2.10.0"
pandera = "^0.16.1"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
pre-commit = "^3.3.3"
black = "^23.7.0"
mypy = "^1.4.1"
pandas-stubs = "^2.0.2.230605"
types-google-cloud-ndb = "^2.1.0.8"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
